requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
PyMySQL==1.1.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
rapidfuzz==3.6.1
slowapi==0.1.9
pydantic-settings==2.1.0